refactor: move lose views into their own folders
